+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for NeoFS Node
 ### Fixed
 - Session v2 token was not supported in the new container + eACL RPC (#4056)
 - Full reads from compressed combined FSTree objects corrupting the zstd stream (#4055)
+- Wrong (1028) status code returned for missing session token verb, should be 2048 (#4052)
 
 ### Changed
 - Optimized GET for fully buffered payloads (#4058)